refactor(scan): centralize full exclude handling
